Report JsonWriters
